+++ app/src/main/java/com/aamo/cookbook/features/recipe/list/screens/RecipesByCategoryScreen.kt
@@ -46,9 +46,9 @@ import com.aamo.cookbook.database.entities.Recipe
 import com.aamo.cookbook.features.recipe.list.components.RecipeCard
 import com.aamo.cookbook.features.recipe.list.models.RecipeListRecipeModel
 import com.aamo.cookbook.features.recipe.list.use_cases.fetchRecipes
-import com.aamo.cookbook.ui.components.LoadingScreen
-import com.aamo.cookbook.ui.components.PrimaryTopAppBar
 import com.aamo.cookbook.ui.theme.CookbookTheme
+import com.aamo.cookbook.utility.components.LoadingScreen
+import com.aamo.cookbook.utility.components.PrimaryTopAppBar
 import com.aamo.cookbook.utility.extensions.general.EMPTY
 import com.aamo.cookbook.utility.extensions.general.ifElse
 import com.aamo.cookbook.utility.tags.UITag
@@ -101,7 +101,7 @@ class RecipesByCategoryScreenViewModel(
 }
 
 fun NavGraphBuilder.recipesByCategoryScreen(
-  onOpenRecipe: (id: Int) -> Unit,
+  onOpenRecipe: (id: Long) -> Unit,
   onOpenSearch: () -> Unit,
   onOpenRecipeForm: () -> Unit,
   onBack: () -> Unit
